mock_api/src/pyproject.toml
[project]
name = "mock-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Flávio Henrique Bonfim",email = "flaviohbonfim@live.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "faker (>=37.1.0,<38.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
